requirements.txt
openai>=1.51.2
python-dotenv==1.0.1
Telethon==1.37.0
PyYAML==6.0.2
python-dotenv==1.0.1
pydantic==2.9.2
